useHistory when click budget on dropdown for peristing pages on rerender
